[SPARK-31748][PYTHON][3.0] Document resource module in PySpark doc and rename/move classes
